refactor: rename package and folder 'logger' -> 'log'.
