+++ src/infrastructure/log/print_logger.go
@@ -1,4 +1,4 @@
-package logger
+package log
 
 import (
 	"fmt"
