+++ src/tests/usecase/create_driver_test.go
@@ -8,13 +8,13 @@ import (
 
 	repository "github.com/rafaft/truck-driver-trip-system/adapter/gateway/database"
 	"github.com/rafaft/truck-driver-trip-system/entity"
-	"github.com/rafaft/truck-driver-trip-system/infrastructure/logger"
+	"github.com/rafaft/truck-driver-trip-system/infrastructure/log"
 	"github.com/rafaft/truck-driver-trip-system/usecase"
 )
 
 func TestCreateDriver(t *testing.T) {
 	now := time.Now()
-	l := logger.NewFakeLogger()
+	l := log.NewFakeLogger()
 	r := repository.NewDriverInMemory()
 	uc := usecase.NewCreateDriver(l, r)
